Move handles to their own test
