+++ crates/test-rust-wasm/src/bin/handles.rs
@@ -1,0 +1,3 @@
+include!("../../../../tests/runtime/handles/wasm.rs");
+
+fn main() {}
