+++ tests/runtime/imports.c
@@ -131,126 +131,6 @@ static void test_flavorful() {
   }
 }
 
-static void test_handles() {
-  host_host_state_t s = host_host_state_create();
-  assert(host_host_state_get(s) == 100);
-  host_host_state_free(&s);
-
-  assert(host_host_state2_saw_close() == false);
-  host_host_state2_t s2 = host_host_state2_create();
-  assert(host_host_state2_saw_close() == false);
-  host_host_state2_free(&s2);
-  assert(host_host_state2_saw_close() == true);
-
-  {
-    host_host_state_t a, b;
-    host_host_state2_t c, d;
-
-    a = host_host_state_create();
-    c = host_host_state2_create();
-    host_two_host_states(a, c, &b, &d);
-    host_host_state_free(&a);
-    host_host_state_free(&b);
-    host_host_state2_free(&c);
-
-    {
-      host_host_state_param_record_t a;
-      a.a = d;
-      host_host_state2_param_record(&a);
-    }
-    {
-      host_host_state_param_tuple_t a;
-      a.f0 = d;
-      host_host_state2_param_tuple(&a);
-    }
-    {
-      host_host_state_param_option_t a;
-      a.tag = 1;
-      a.val = d;
-      host_host_state2_param_option(&a);
-    }
-    {
-      host_host_state_param_result_t a;
-      a.tag = 0;
-      a.val.ok = d;
-      host_host_state2_param_result(&a);
-      a.tag = 1;
-      a.val.err = 2;
-      host_host_state2_param_result(&a);
-    }
-    {
-      host_host_state_param_variant_t a;
-      a.tag = HOST_HOST_STATE_PARAM_VARIANT_0;
-      a.val.f0 = d;
-      host_host_state2_param_variant(&a);
-      a.tag = HOST_HOST_STATE_PARAM_VARIANT_1;
-      a.val.f1 = 2;
-      host_host_state2_param_variant(&a);
-    }
-    {
-      host_host_state2_t arr[2];
-      arr[0] = d;
-      arr[1] = d;
-      host_list_host_state2_t list;
-      list.len = 0;
-      list.ptr = arr;
-      host_host_state2_param_list(&list);
-      list.len = 1;
-      host_host_state2_param_list(&list);
-      list.len = 2;
-      host_host_state2_param_list(&list);
-    }
-
-    host_host_state2_free(&d);
-  }
-
-  {
-    host_host_state_result_record_t a;
-    host_host_state2_result_record(&a);
-    host_host_state2_free(&a.a);
-  }
-  {
-    host_host_state2_t a;
-    host_host_state2_result_tuple(&a);
-    host_host_state2_free(&a);
-  }
-  {
-    host_host_state2_t a;
-    assert(host_host_state2_result_option(&a));
-    host_host_state2_free(&a);
-  }
-  {
-    host_host_state_result_result_t a;
-    host_host_state2_result_result(&a);
-    assert(a.tag == 0);
-    host_host_state2_free(&a.val.ok);
-  }
-  {
-    host_host_state_result_variant_t a;
-    host_host_state2_result_variant(&a);
-    assert(a.tag == 0);
-    host_host_state2_free(&a.val.f0);
-  }
-  {
-    host_list_host_state2_t a;
-    host_host_state2_result_list(&a);
-    host_list_host_state2_free(&a);
-  }
-  {
-    host_markdown2_t a = host_markdown2_create();
-    host_string_t s;
-    host_string_set(&s, "red is the best color");
-    host_markdown2_append(a, &s);
-    host_markdown2_render(a, &s);
-
-    const char *expected = "green is the best color";
-    assert(s.len == strlen(expected));
-    assert(memcmp(s.ptr, expected, s.len) == 0);
-    host_string_free(&s);
-    host_markdown2_free(&a);
-  }
-}
-
 static void test_buffers() {
   {
     host_push_buffer_u8_t push;
